refactor: extract nav links into component
